+++ examples/dynamic_behavior/skip_messages.cpp
@@ -2,32 +2,46 @@
 // messages in the mailbox until an actor switches to another behavior that can
 // handle them.
 
+#include "caf/actor_from_state.hpp"
 #include "caf/actor_ostream.hpp"
 #include "caf/actor_system.hpp"
 #include "caf/caf_main.hpp"
 #include "caf/event_based_actor.hpp"
+#include "caf/fwd.hpp"
+#include "caf/mail_cache.hpp"
 #include "caf/scoped_actor.hpp"
 
 using namespace caf;
 using namespace std::literals;
 
-behavior server(event_based_actor* self) {
-  // Setting the default handler to `skip` will keep all unmatched messages in
-  // the mailbox until we switch to another behavior that handles them. This
-  // allows us to receive idle and ping messages in any order but process them
-  // in alternating order.
-  self->set_default_handler(skip);
-  return {
-    [self](idle_atom, const actor& worker) {
-      // Passing keep_behavior allows us to return to this behavior later on by
-      // calling `self->unbecome()`.
-      self->become(keep_behavior, [self, worker](ping_atom atm) {
-        self->delegate(worker, atm);
-        self->unbecome();
-      });
-    },
-  };
-}
+struct server_state {
+  server_state(event_based_actor* selfptr) : self(selfptr), cache(self, 10) {
+    // nop
+  }
+
+  behavior make_behavior() {
+    return {
+      [this](idle_atom, const actor& worker) {
+        // Passing keep_behavior allows us to return to this behavior later on
+        // by calling `self->unbecome()`.
+        self->become(keep_behavior, [this, worker](ping_atom) {
+          // Switch back to our default behavior, put all messages from the
+          // cache back to the mailbox and delegate 'ping' to the worker.
+          self->unbecome();
+          cache.unstash();
+          return self->mail(ping_atom_v).delegate(worker);
+        });
+      },
+      [this](message msg) {
+        // Stash all messages until we receive an `idle_atom`.
+        cache.stash(std::move(msg));
+      },
+    };
+  }
+
+  event_based_actor* self;
+  mail_cache cache;
+};
 
 behavior client(event_based_actor* self, const actor& serv) {
   self->link_to(serv);
@@ -41,7 +55,7 @@ behavior client(event_based_actor* self, const actor& serv) {
 }
 
 void caf_main(actor_system& sys) {
-  auto serv = sys.spawn(server);
+  auto serv = sys.spawn(actor_from_state<server_state>);
   auto worker = sys.spawn(client, serv);
   scoped_actor self{sys};
   self->mail(ping_atom_v)